remove unnecessary old files and move file_helper to utils package
